breaking down of PropertiesFilePass into 2 separate passes
