+++ src/main/scala/ai/privado/languageEngine/java/passes/config/PropertiesFilePass.scala
@@ -24,65 +24,39 @@
 package ai.privado.languageEngine.java.passes.config
 
 import ai.privado.cache.RuleCache
-import ai.privado.utility.Utilities
-import com.github.wnameless.json.flattener.JsonFlattener
-import io.circe.yaml.parser
-import io.joern.x2cpg.SourceFiles
-import io.shiftleft.codepropertygraph.generated.nodes.{Literal, MethodParameterIn, NewFile, NewJavaProperty}
+import io.shiftleft.codepropertygraph.generated.nodes.{
+  AnnotationParameterAssign,
+  JavaProperty,
+  Literal,
+  Member,
+  MethodParameterIn,
+  NewFile,
+  NewJavaProperty
+}
 import io.shiftleft.codepropertygraph.generated.{Cpg, EdgeTypes}
 import io.shiftleft.passes.ForkJoinParallelCpgPass
 import io.shiftleft.semanticcpg.language._
 import org.slf4j.LoggerFactory
 import overflowdb.BatchedUpdate
 import overflowdb.traversal._
-
-import java.util.Properties
-import scala.jdk.CollectionConverters._
-import scala.util.{Failure, Success, Try}
-import scala.xml._
+import ai.privado.languageEngine.java.language.NodeStarters
 
 /** This pass creates a graph layer for Java `.properties` files.
   */
-class PropertiesFilePass(cpg: Cpg, projectRoot: String, ruleCache: RuleCache)
-    extends ForkJoinParallelCpgPass[String](cpg) {
+class PropertiesFilePass(cpg: Cpg) extends ForkJoinParallelCpgPass[JavaProperty](cpg) {
+  override def generateParts(): Array[_ <: AnyRef] =
+    cpg.property.l.toArray.filter(pair => pair.name.nonEmpty && pair.value.nonEmpty)
 
-  private val logger = LoggerFactory.getLogger(getClass)
-
-  override def generateParts(): Array[String] =
-    propertiesFiles(projectRoot, Set(".properties", ".yml", ".yaml", ".xml")).toArray
-
-  override def runOnPart(builder: DiffGraphBuilder, file: String): Unit = {
-    val fileNode      = addFileNode(file, builder)
-    val propertyNodes = addPropertyNodesAndConnectToUsers(file, builder)
-    propertyNodes.foreach(builder.addEdge(_, fileNode, EdgeTypes.SOURCE_FILE))
+  override def runOnPart(builder: DiffGraphBuilder, property: JavaProperty): Unit = {
+    connectProperties(property, builder)
   }
 
-  private def addPropertyNodesAndConnectToUsers(
-    file: String,
-    builder: BatchedUpdate.DiffGraphBuilder
-  ): List[NewJavaProperty] = {
-    Try {
-      obtainKeyValuePairs(file, builder)
-    } match {
-      case Success(keyValuePairs) =>
-        val propertyNodes = keyValuePairs.map(addPropertyNode(_, builder))
-
-        propertyNodes.foreach(propertyNode => {
-          connectGetPropertyLiterals(propertyNode, builder)
-          connectAnnotatedParameters(propertyNode, builder)
-        })
-
-        propertyNodes
-      case Failure(exception) =>
-        logger.warn(exception.getMessage)
-        List()
-    }
+  private def connectProperties(property: JavaProperty, builder: DiffGraphBuilder): Unit = {
+    connectGetPropertyLiterals(property, builder)
+    connectAnnotatedParameters(property, builder)
   }
 
-  private def connectAnnotatedParameters(
-    propertyNode: NewJavaProperty,
-    builder: BatchedUpdate.DiffGraphBuilder
-  ): Unit = {
+  private def connectAnnotatedParameters(propertyNode: JavaProperty, builder: BatchedUpdate.DiffGraphBuilder): Unit = {
     val paramsAndValues = annotatedParameters()
 
     paramsAndValues
@@ -117,21 +91,16 @@ class PropertiesFilePass(cpg: Cpg, projectRoot: String, ruleCache: RuleCache)
 
   /** List of all members annotated with Spring's `Value` annotation, along with the property name.
     */
-  private def annotatedMembers() = cpg.annotation
+  private def annotatedMembers(): List[(AnnotationParameterAssign, Member)] = cpg.annotation
     .fullName(".*Value.*")
     .where(_.member)
     .filter(_.parameterAssign.l.length > 0)
     .map { x => (x.parameterAssign.head, x.member.head) }
     .l
-  private def getMember(member: String, className: String) =
-    cpg.member.where(_.typeDecl.fullName(className)).where(_.name(member)).toList
 
   /** In this method, we attempt to identify users of properties and connect them to property nodes.
     */
-  private def connectGetPropertyLiterals(
-    propertyNode: NewJavaProperty,
-    builder: BatchedUpdate.DiffGraphBuilder
-  ): Unit = {
+  private def connectGetPropertyLiterals(propertyNode: JavaProperty, builder: BatchedUpdate.DiffGraphBuilder): Unit = {
     matchingLiteralsInGetPropertyCalls(propertyNode.name).foreach { lit =>
       builder.addEdge(propertyNode, lit, EdgeTypes.IS_USED_AT)
       builder.addEdge(lit, propertyNode, EdgeTypes.ORIGINAL_PROPERTY)
@@ -142,158 +111,5 @@ class PropertiesFilePass(cpg: Cpg, projectRoot: String, ruleCache: RuleCache)
     .codeExact("\"" + propertyName + "\"")
     .where(_.inCall.name(".*getProperty"))
     .l
-
-  private def obtainKeyValuePairs(file: String, builder: DiffGraphBuilder): List[(String, String)] = {
-    if (file.matches(""".*\.(?:yml|yaml)""")) {
-      loadAndConvertYMLtoProperties(file)
-    } else if (file.endsWith(".xml")) {
-      loadAndConvertXMLtoProperties(file, builder)
-    } else {
-      loadFromProperties(file)
-    }
-  }
-
-  private def loadFromProperties(file: String): List[(String, String)] = {
-    val properties  = new Properties()
-    val inputStream = better.files.File(file).newFileInputStream
-    properties.load(inputStream)
-    inputStream.close()
-    propertiesToKeyValuePairs(properties)
-  }
-
-  private def loadAndConvertYMLtoProperties(file: String): List[(String, String)] = {
-    parser.parse(better.files.File(file).contentAsString) match {
-      case Right(json) => {
-        try {
-          JsonFlattener
-            .flattenAsMap(json.toString)
-            .asScala
-            .toList
-            .collect(p => (p._1, p._2.toString))
-            .toList
-        } catch {
-          case e: Throwable =>
-            logger.trace(s"Error while creating properties node for file $file")
-            logger.debug(s"Error while creating properties node for file : $file, error : ${e.getMessage}")
-            List[("", "")]()
-        }
-      }
-      case Left(error) => {
-        List[("", "")]()
-      }
-    }
-  }
-
-  // if beans file contain placeholders, search in .properties files across the project
-  private def resolvePlaceholderValuesXML(placeholder: String): String = {
-    val propertyFiles: List[String] = propertiesFiles(projectRoot, Set(".properties"))
-    propertyFiles.foreach(file => {
-      // Search across properties to find the required
-      loadFromProperties(file).foreach(propertyValue => {
-        val (name, value) = propertyValue;
-        if (name.equals(placeholder)) return value;
-      })
-    })
-    ""
-  }
-
-  // Used to extract (name, value) pairs from a bean config file
-  private def XMLParserBean(xmlPath: String, builder: DiffGraphBuilder): List[(String, String)] = {
-    try {
-      val xml = XML.loadFile(xmlPath)
-      val nameValuePairs = (xml \\ "bean").flatMap { bean =>
-        {
-          var result: (String, String) = ("", "")
-          val className: String        = bean \@ "class"
-          (bean \\ "property").map { prop =>
-            {
-              // Search for property tags inside a bean
-              val propValue = prop \@ "value"
-              if (propValue.startsWith("$") && propValue.endsWith("}")) {
-                val value = resolvePlaceholderValuesXML(
-                  propValue.substring(2, propValue.length - 1)
-                ) // Pass placeholder name without ${ and }
-                if (value.nonEmpty) {
-                  result = ((prop \@ "name"), value)
-                }
-              } else {
-                result = ((prop \@ "name"), propValue)
-              }
-
-            }
-
-            val members = getMember((prop \@ "name"), className);
-            if (members.nonEmpty) {
-              val propertyNode = NewJavaProperty().name(result._1).value(result._2)
-              val member       = members.head
-              if (member != null) {
-                builder.addEdge(propertyNode, member, EdgeTypes.IS_USED_AT)
-                builder.addEdge(member, propertyNode, EdgeTypes.ORIGINAL_PROPERTY);
-              }
-            }
-            result
-          }
-        }
-      }
-
-      return nameValuePairs.toList
-        .collect { case (name, value) => if (value.nonEmpty) (name, value) else ("", "") }
-        .filter { case (name, value) =>
-          name.nonEmpty && value.nonEmpty // Filter out name, value pairs which could not be resolved
-        }
-    } catch {
-      case e: Throwable => println(e)
-    }
-
-    List[("", "")]()
-
-  }
-
-  private def loadAndConvertXMLtoProperties(file: String, builder: DiffGraphBuilder): List[(String, String)] = {
-    val properties  = new Properties();
-    val inputStream = better.files.File(file).newInputStream
-    try {
-      properties.loadFromXML(inputStream)
-      properties.propertyNames.asScala.toList
-        .collect(p => (p.toString, properties.getProperty(p.toString)))
-    } catch {
-      case _: Throwable => {
-        XMLParserBean(file, builder)
-      }
-    }
-  }
-
-  private def propertiesToKeyValuePairs(properties: Properties): List[(String, String)] = {
-    properties
-      .propertyNames()
-      .asScala
-      .collect { case key: String =>
-        (key, properties.getProperty(key))
-      }
-      .toList
-  }
-
-  // Add extensions as a parameter to decouple it
-  private def propertiesFiles(projectRoot: String, extensions: Set[String]): List[String] = {
-    SourceFiles
-      .determine(Set(projectRoot), extensions)
-      .filter(Utilities.isFileProcessable(_, ruleCache))
-  }
-
-  private def addFileNode(name: String, builder: BatchedUpdate.DiffGraphBuilder): NewFile = {
-    val fileNode = NewFile().name(name)
-    builder.addNode(fileNode)
-    fileNode
-  }
-
-  private def addPropertyNode(
-    keyValuePair: (String, String),
-    builder: BatchedUpdate.DiffGraphBuilder
-  ): NewJavaProperty = {
-    val (key, value) = keyValuePair
-    val propertyNode = NewJavaProperty().name(key).value(value)
-    builder.addNode(propertyNode)
-    propertyNode
-  }
 
 }
