+++ src/test/scala/ai/privado/languageEngine/java/passes/config/PropertiesFilePassTest.scala
@@ -26,6 +26,8 @@ package ai.privado.languageEngine.java.passes.config
 import ai.privado.cache.RuleCache
 import ai.privado.languageEngine.java.language._
 import ai.privado.languageEngine.python.passes.config.PythonPropertyFilePass
+import ai.privado.model.Language
+import ai.privado.utility.PropertyCollectorPass
 import better.files.File
 import io.joern.console.cpgcreation.PythonSrcCpgGenerator
 import io.shiftleft.codepropertygraph.generated.Cpg
@@ -251,7 +253,8 @@ abstract class PropertiesFilePassTestBase(fileExtension: String)
     val config = Config(inputPath = inputDir.toString(), outputPath = outputFile.toString())
 
     cpg = new JavaSrc2Cpg().createCpg(config).get
-    new PropertiesFilePass(cpg, inputDir.toString, new RuleCache()).createAndApply()
+    new PropertyCollectorPass(cpg, inputDir.toString(), new RuleCache, Language.JAVA).createAndApply()
+    new PropertiesFilePass(cpg).createAndApply()
 
     super.beforeAll()
   }
